wado-compiler: extract lower submodule from cm_binding
Move the memory-based CM lower (synthesize_lower, the tuple/option/
variant variants, and synthesize_lower_wasi_type_to_memory) and the
flat-args flattener (synthesize_flatten_value_to_flat_args and the
Option<T> wrapper) into `synthesis/cm_binding/lower.rs`. The dual
shapes (memory-based and stack-based) live together because they
share the WASI variant / option logic and are paired with a single
high-level lowering decision in the import adapter.
